Move FlowNetSD to flownet_sd
