+++ respiration/extractor/flownet/__init__.py
@@ -1,2 +1,2 @@
-from .flownet_2_sd import FlowNet2SD
+from .flownet_sd import FlowNet2SD
 from .preprocessing import *
